Print now uses "int" as coordinates because it looked strange when using values like 0.5.
git-svn-id: file:///private-backup/var/svn/jngl@66 8b392cba-0c28-4fbc-ad53-5d9c93102ef3

+++ jngl.hpp
@@ -152,8 +152,8 @@ namespace jngl
 				        unsigned char alpha = 255);
 
 	void Print(const std::string& text,
-	           double xposition,
-	           double yposition);
+	           int xposition,
+	           int yposition);
 
 	void SetFontSize(int size);
